[py3] Cast all division results back to int
When connecting to Live 11.2.6 with my Launchpad Pro mk1, the sequencer throws
errors related to unexpected float types for several of the side buttons (eg.
change octave, enter melodic mode).

This commit attempts to find all places where division is used and a floating
point number is returned instead of an int (Python 3 changes the default
behavior of int division).

I've tested the step sequencer mode and all features seem to work. I'm unable to
reproduce any of the errors in the logs now.

+++ NoteEditorComponent.py
@@ -128,7 +128,7 @@ class NoteEditorComponent(ControlSurfaceComponent):
 		if self.is_multinote:
 			self._page = page
 		else:
-			self._page = page / 4  # number of line per note ?
+			self._page = int(page / 4)  # number of line per note ?
 
 	def set_clip(self, clip):
 		self._clip = clip
@@ -161,7 +161,7 @@ class NoteEditorComponent(ControlSurfaceComponent):
 			self._grid_back_buffer[self._page % self.width][i] = "StepSequencer.NoteEditor.PageMarker"
 
 	def _display_note_markers(self):
-		for i in range(0, self.height / self.number_of_lines_per_note):
+		for i in range(0, int(self.height / self.number_of_lines_per_note)):
 			if self._key_index_is_root_note[i]:
 				for j in range(0, self.number_of_lines_per_note):
 					self._grid_back_buffer[0][self.height - i * self.number_of_lines_per_note - j - 1] = "StepSequencer.NoteEditor.NoteMarker"
@@ -331,7 +331,7 @@ class NoteEditorComponent(ControlSurfaceComponent):
 
 				if self.is_multinote:
 					time = self.quantization * (self._page * self.width * self.number_of_lines_per_note + x + (y % self.number_of_lines_per_note * self.width))
-					pitch = self._key_indexes[8 / self.number_of_lines_per_note - 1 - y / self.number_of_lines_per_note]
+					pitch = self._key_indexes[int(8 / self.number_of_lines_per_note - 1 - y / self.number_of_lines_per_note)]
 				else:
 					time = self.quantization * (self._page * self.width * self.number_of_lines_per_note + y * self.width + x)
 					pitch = self._key_indexes[0]
